requirements.txt
aiofiles==25.1.0
beautifulsoup4==4.14.2
fastapi==0.121.3
uvicorn                   
matplotlib==3.10.7
numpy==2.3.5
ollama==0.6.1
openpyxl==3.1.5
pandas==2.3.3
pyaxis==0.4.1
pydantic==2.12.4
python-dotenv==1.2.1
requests==2.32.5
SQLAlchemy==2.0.44
psycopg2-binary          
thefuzz==0.22.1
playwright==1.53.0